+++ tests/test_dtool_dataset_freeze.py
@@ -40,3 +40,25 @@ def test_dataset_freeze_functional(chdir_fixture):  # NOQA
 
     # Manifest has been updated.
     assert len(dataset.identifiers) == 1
+
+
+def test_dataset_freeze_rogue_content_functional(chdir_fixture):  # NOQA
+    from dtool_create.dataset import create, freeze
+    runner = CliRunner()
+
+    dataset_name = "my_dataset"
+    result = runner.invoke(create, [dataset_name])
+    assert result.exit_code == 0
+
+    # At this point we have a proto dataset
+    dataset_abspath = os.path.abspath(dataset_name)
+    dataset_uri = "file://{}".format(dataset_abspath)
+
+    # Create rogue data file
+    sample_file_name = os.path.join(dataset_abspath, "hello.txt")
+    with open(sample_file_name, "w") as fh:
+        fh.write("hello world")
+
+    result = runner.invoke(freeze, [dataset_uri])
+    assert result.exit_code == 4
+    assert result.output.find("Rogue content") != -1